requirements.txt
openai==0.27.0
tweepy==4.14.0
python-telegram-bot==13.15
